ipf: Define offsetof() more conventionally, fix pointer cast warnings

+++ usr/src/uts/common/inet/ipf/netinet/ip_compat.h
@@ -1759,7 +1759,7 @@ typedef	struct	tcpiphdr	tcpiphdr_t;
 # undef	offsetof
 #endif
 #ifndef offsetof
-# define offsetof(t,m) (int)((&((t *)0L)->m))
+# define offsetof(t,m) (size_t)((&((t *)0)->m))
 #endif
 
 /*
